requirements.txt
psycopg2-binary==2.9.9
snowflake-connector-python==3.7.0
hvac==2.1.0
boto3==1.34.0
botocore==1.34.0
requests==2.31.0

